w3c-reports/12092014/mac-wk2/runned-tests/2dcontext/text-styles/2d.text.draw.space.collapse.other-expected.txt

2d.text.draw.space.collapse.other
Space characters are converted to U+0020, and collapsed (per CSS)
Actual output:
Expected output:


FAIL Space characters are converted to U+0020, and collapsed (per CSS) assert_approx_equals: Red channel of the pixel at (25, 25) expected 0 +/- 2 but got 255(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_approx_equals@http://localhost:8800/resources/testharness.js:425:15
_assertPixelApprox@http://localhost:8800/common/canvas-tests.js:47:25
http://localhost:8800/2dcontext/text-styles/2d.text.draw.space.collapse.other.html:36:23
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:843:37)
